multiply ADC readings before dividing for average to preserve addition precision obtained in averaging
git-svn-id: tumanako.svn.sourceforge.net::svn/tumanako/trunk@411 c3429617-7ab9-49b5-8233-5f90b2ce515f

+++ src/c/util.h
@@ -35,7 +35,7 @@ void sleep(unsigned char time);
 void writeEEPROM(unsigned char address, unsigned char value);
 unsigned char readEEPROM(unsigned char address);
 
-unsigned short adc(unsigned char c);
+unsigned long adc(unsigned char c);
 
 void restoreLed();
